workflows/react_distributed/evaluator/requirements.txt
bedrock-agentcore
boto3>=1.40.0
langchain
langgraph
langchain-openai
python-dotenv
